requirements.txt
Django==3.2.13
djangorestframework==3.12.4
psycopg2-binary==2.8.6
python-dotenv==0.19.2
django-filter==2.4.0
drf-yasg==1.20.0
gunicorn==20.0.4
django-debug-toolbar==3.2.4